+++ docs/source/conf.py
@@ -20,8 +20,8 @@ sys.path.insert(0, os.path.abspath('../../'))
 # -- Project information -----------------------------------------------------
 
 project = 'probatus'
-copyright = '2019, RPAA'
-author = 'RPAA'
+copyright = '2020, ING Bank N.V.'
+author = 'ING Bank N.V.'
 
 # The short X.Y version
 version = ''
@@ -44,6 +44,10 @@ extensions = [
     'sphinx.ext.todo',
     'sphinx.ext.mathjax',
     'sphinx.ext.githubpages',
+    'sphinx.ext.intersphinx',
+    'sphinx.ext.coverage',
+    'sphinx.ext.ifconfig',
+    'sphinx.ext.viewcode',
     'nbsphinx'
 ]
 
@@ -139,7 +143,7 @@ latex_elements = {
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
     (master_doc, 'probatus.tex', 'probatus Documentation',
-     'RPAA', 'manual'),
+     'ING Bank N.V.', 'manual'),
 ]