+++ probatus/interpret/inspector.py
@@ -1,10 +1,30 @@
+# Copyright (c) 2020 ING Bank N.V.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy of
+# this software and associated documentation files (the "Software"), to deal in
+# the Software without restriction, including without limitation the rights to
+# use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+# the Software, and to permit persons to whom the Software is furnished to do so,
+# subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+# CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+
+
 from ..utils import NotFittedError, UnsupportedModelError
 import numpy as np
 import pandas as pd
 import copy
 
 from sklearn.cluster import KMeans
-from probatus.interpret import _shap_helpers
+from probatus.utils import shap_helpers
 
 
 def return_confusion_metric(y_true, y_score, normalize = False):
@@ -259,7 +279,7 @@ class InspectorShap(BaseInspector):
         predicted_proba = pd.Series(self.compute_probabilities(X), index = y.index,name = 'pred_proba')
 
         # Compute SHAP values and cluster them
-        X_shap = _shap_helpers.shap_to_df(self.model, X, **shap_kwargs)
+        X_shap = shap_helpers.shap_to_df(self.model, X, **shap_kwargs)
         if fit_clusters:
             self.fit_clusters(X_shap)
         clusters = pd.Series(self.predict_clusters(X_shap), index=y.index, name='cluster_id')