+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2020-09-30
+- Add resemblance model, with SHAP based importance
+- Improve the docs for resemblance model
+- Refactor stats tests, improve docs and expose functionality to users
+
+## [1.1.1] - 2020-09-08
+- Improve Tree Bucketer, enable user to pass own tree object
+
+## [1.1.0] - 2020-08-24
+- Improve docs for stats_tests
+- Refactor stats_tests
+
 ## [1.0.1] - 2020-08-07
 - TreeBucketer, which bins the data based on the target distribution, using Decision Trees fitted on a single feature
 - PSI calculation includes the p-values calculation